add helper functions
