+++ crates/cheatnet/src/runtime_extensions/forge_runtime_extension/cheatcodes/cheat_execution_info.rs
@@ -1,9 +1,9 @@
-use super::spoof::ResourceBounds;
 use crate::{
     state::{CheatSpan, CheatStatus},
     CheatnetState,
 };
 use cairo_felt::Felt252;
+use conversions::felt252::SerializeAsFelt252Vec;
 use runtime::FromReader;
 use starknet_api::core::{ContractAddress, EntryPointSelector};
 
@@ -22,6 +22,21 @@ pub enum Operation<T> {
     StopGlobal,
     #[default]
     Retain,
+}
+
+#[derive(FromReader, Clone, Default, Debug, Eq, PartialEq)]
+pub struct ResourceBounds {
+    pub resource: Felt252,
+    pub max_amount: u64,
+    pub max_price_per_unit: u128,
+}
+
+impl SerializeAsFelt252Vec for ResourceBounds {
+    fn serialize_into_felt252_vec(self, output: &mut Vec<Felt252>) {
+        output.push(self.resource);
+        output.push(self.max_amount.into());
+        output.push(self.max_price_per_unit.into());
+    }
 }
 
 #[derive(Clone, Default, Debug, PartialEq, Eq)]
