+++ crates/cheatnet/src/runtime_extensions/forge_runtime_extension/cheatcodes/spoof.rs
@@ -1,29 +1,12 @@
 use super::cheat_execution_info::{
-    CheatArguments, ExecutionInfoMockOperations, Operation, TxInfoMockOperations,
+    CheatArguments, ExecutionInfoMockOperations, Operation, ResourceBounds, TxInfoMockOperations,
 };
 use crate::state::CheatSpan;
 use crate::CheatnetState;
 use cairo_felt::Felt252;
-use conversions::felt252::SerializeAsFelt252Vec;
-use runtime::FromReader;
 use starknet_api::core::ContractAddress;
 
-#[derive(FromReader, Clone, Default, Debug, Eq, PartialEq)]
-pub struct ResourceBounds {
-    pub resource: Felt252,
-    pub max_amount: u64,
-    pub max_price_per_unit: u128,
-}
-
-impl SerializeAsFelt252Vec for ResourceBounds {
-    fn serialize_into_felt252_vec(self, output: &mut Vec<Felt252>) {
-        output.push(self.resource);
-        output.push(self.max_amount.into());
-        output.push(self.max_price_per_unit.into());
-    }
-}
-
-#[derive(FromReader, Clone, Default, Debug)]
+#[derive(Clone, Default, Debug)]
 pub struct TxInfoMock {
     pub version: Option<Felt252>,
     pub account_contract_address: Option<Felt252>,
