+++ snforge_std/src/lib.cairo
@@ -5,10 +5,6 @@ use cheatcodes::contract_class::get_class_hash;
 use cheatcodes::contract_class::ContractClass;
 use cheatcodes::contract_class::ContractClassTrait;
 
-use cheatcodes::tx_info::spoof;
-use cheatcodes::tx_info::start_spoof;
-use cheatcodes::tx_info::stop_spoof;
-
 use cheatcodes::l1_handler::L1Handler;
 use cheatcodes::l1_handler::L1HandlerTrait;
 
@@ -30,26 +26,6 @@ use cheatcodes::ContractAddress;
 use cheatcodes::CheatSpan;
 use cheatcodes::test_address;
 use cheatcodes::test_selector;
-use cheatcodes::cheat_caller_address;
-use cheatcodes::cheat_caller_address_global;
-use cheatcodes::start_cheat_caller_address;
-use cheatcodes::stop_cheat_caller_address;
-use cheatcodes::stop_cheat_caller_address_global;
-use cheatcodes::cheat_block_number;
-use cheatcodes::cheat_block_number_global;
-use cheatcodes::start_cheat_block_number;
-use cheatcodes::stop_cheat_block_number;
-use cheatcodes::stop_cheat_block_number_global;
-use cheatcodes::cheat_block_timestamp;
-use cheatcodes::cheat_block_timestamp_global;
-use cheatcodes::start_cheat_block_timestamp;
-use cheatcodes::stop_cheat_block_timestamp;
-use cheatcodes::stop_cheat_block_timestamp_global;
-use cheatcodes::cheat_sequencer_address;
-use cheatcodes::cheat_sequencer_address_global;
-use cheatcodes::start_cheat_sequencer_address;
-use cheatcodes::stop_cheat_sequencer_address;
-use cheatcodes::stop_cheat_sequencer_address_global;
 use cheatcodes::mock_call;
 use cheatcodes::start_mock_call;
 use cheatcodes::stop_mock_call;
@@ -60,6 +36,88 @@ use cheatcodes::execution_info::BlockInfoMockImpl;
 use cheatcodes::execution_info::TxInfoMock;
 use cheatcodes::execution_info::Operation;
 use cheatcodes::execution_info::CheatArguments;
+
+use cheatcodes::execution_info::caller_address::cheat_caller_address;
+use cheatcodes::execution_info::caller_address::cheat_caller_address_global;
+use cheatcodes::execution_info::caller_address::stop_cheat_caller_address;
+use cheatcodes::execution_info::caller_address::stop_cheat_caller_address_global;
+use cheatcodes::execution_info::caller_address::start_cheat_caller_address;
+use cheatcodes::execution_info::block_number::cheat_block_number;
+use cheatcodes::execution_info::block_number::cheat_block_number_global;
+use cheatcodes::execution_info::block_number::stop_cheat_block_number;
+use cheatcodes::execution_info::block_number::stop_cheat_block_number_global;
+use cheatcodes::execution_info::block_number::start_cheat_block_number;
+use cheatcodes::execution_info::block_timestamp::cheat_block_timestamp;
+use cheatcodes::execution_info::block_timestamp::cheat_block_timestamp_global;
+use cheatcodes::execution_info::block_timestamp::stop_cheat_block_timestamp;
+use cheatcodes::execution_info::block_timestamp::stop_cheat_block_timestamp_global;
+use cheatcodes::execution_info::block_timestamp::start_cheat_block_timestamp;
+use cheatcodes::execution_info::sequencer_address::cheat_sequencer_address;
+use cheatcodes::execution_info::sequencer_address::cheat_sequencer_address_global;
+use cheatcodes::execution_info::sequencer_address::stop_cheat_sequencer_address;
+use cheatcodes::execution_info::sequencer_address::stop_cheat_sequencer_address_global;
+use cheatcodes::execution_info::sequencer_address::start_cheat_sequencer_address;
+use cheatcodes::execution_info::version::cheat_version;
+use cheatcodes::execution_info::version::cheat_version_global;
+use cheatcodes::execution_info::version::stop_cheat_version;
+use cheatcodes::execution_info::version::stop_cheat_version_global;
+use cheatcodes::execution_info::version::start_cheat_version;
+use cheatcodes::execution_info::max_fee::cheat_max_fee;
+use cheatcodes::execution_info::max_fee::cheat_max_fee_global;
+use cheatcodes::execution_info::max_fee::stop_cheat_max_fee;
+use cheatcodes::execution_info::max_fee::stop_cheat_max_fee_global;
+use cheatcodes::execution_info::max_fee::start_cheat_max_fee;
+use cheatcodes::execution_info::signature::cheat_signature;
+use cheatcodes::execution_info::signature::cheat_signature_global;
+use cheatcodes::execution_info::signature::stop_cheat_signature;
+use cheatcodes::execution_info::signature::stop_cheat_signature_global;
+use cheatcodes::execution_info::signature::start_cheat_signature;
+use cheatcodes::execution_info::transaction_hash::cheat_transaction_hash;
+use cheatcodes::execution_info::transaction_hash::cheat_transaction_hash_global;
+use cheatcodes::execution_info::transaction_hash::stop_cheat_transaction_hash;
+use cheatcodes::execution_info::transaction_hash::stop_cheat_transaction_hash_global;
+use cheatcodes::execution_info::transaction_hash::start_cheat_transaction_hash;
+use cheatcodes::execution_info::chain_id::cheat_chain_id;
+use cheatcodes::execution_info::chain_id::cheat_chain_id_global;
+use cheatcodes::execution_info::chain_id::stop_cheat_chain_id;
+use cheatcodes::execution_info::chain_id::stop_cheat_chain_id_global;
+use cheatcodes::execution_info::chain_id::start_cheat_chain_id;
+use cheatcodes::execution_info::nonce::cheat_nonce;
+use cheatcodes::execution_info::nonce::cheat_nonce_global;
+use cheatcodes::execution_info::nonce::stop_cheat_nonce;
+use cheatcodes::execution_info::nonce::stop_cheat_nonce_global;
+use cheatcodes::execution_info::nonce::start_cheat_nonce;
+use cheatcodes::execution_info::resource_bounds::cheat_resource_bounds;
+use cheatcodes::execution_info::resource_bounds::cheat_resource_bounds_global;
+use cheatcodes::execution_info::resource_bounds::stop_cheat_resource_bounds;
+use cheatcodes::execution_info::resource_bounds::stop_cheat_resource_bounds_global;
+use cheatcodes::execution_info::resource_bounds::start_cheat_resource_bounds;
+use cheatcodes::execution_info::tip::cheat_tip;
+use cheatcodes::execution_info::tip::cheat_tip_global;
+use cheatcodes::execution_info::tip::stop_cheat_tip;
+use cheatcodes::execution_info::tip::stop_cheat_tip_global;
+use cheatcodes::execution_info::tip::start_cheat_tip;
+use cheatcodes::execution_info::paymaster_data::cheat_paymaster_data;
+use cheatcodes::execution_info::paymaster_data::cheat_paymaster_data_global;
+use cheatcodes::execution_info::paymaster_data::stop_cheat_paymaster_data;
+use cheatcodes::execution_info::paymaster_data::stop_cheat_paymaster_data_global;
+use cheatcodes::execution_info::paymaster_data::start_cheat_paymaster_data;
+use cheatcodes::execution_info::nonce_data_availability_mode::cheat_nonce_data_availability_mode;
+use cheatcodes::execution_info::nonce_data_availability_mode::cheat_nonce_data_availability_mode_global;
+use cheatcodes::execution_info::nonce_data_availability_mode::stop_cheat_nonce_data_availability_mode;
+use cheatcodes::execution_info::nonce_data_availability_mode::stop_cheat_nonce_data_availability_mode_global;
+use cheatcodes::execution_info::nonce_data_availability_mode::start_cheat_nonce_data_availability_mode;
+use cheatcodes::execution_info::fee_data_availability_mode::cheat_fee_data_availability_mode;
+use cheatcodes::execution_info::fee_data_availability_mode::cheat_fee_data_availability_mode_global;
+use cheatcodes::execution_info::fee_data_availability_mode::stop_cheat_fee_data_availability_mode;
+use cheatcodes::execution_info::fee_data_availability_mode::stop_cheat_fee_data_availability_mode_global;
+use cheatcodes::execution_info::fee_data_availability_mode::start_cheat_fee_data_availability_mode;
+use cheatcodes::execution_info::account_deployment_data::cheat_account_deployment_data;
+use cheatcodes::execution_info::account_deployment_data::cheat_account_deployment_data_global;
+use cheatcodes::execution_info::account_deployment_data::stop_cheat_account_deployment_data;
+use cheatcodes::execution_info::account_deployment_data::stop_cheat_account_deployment_data_global;
+use cheatcodes::execution_info::account_deployment_data::start_cheat_account_deployment_data;
+
 
 mod fs;
